requirements.txt
opencv-python==4.13.0.92
mediapipe==0.10.9
ultralytics
pandas
matplotlib
streamlit
fpdf2
numpy